requirements.txt
alembic==1.11.1
altair==5.0.1
appnope==0.1.3
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cmaes==0.10.0
colorlog==6.7.0
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
deep-translator==1.11.4
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.12.2
fsspec==2023.6.0
git-filter-repo==2.38.0
gitdb==4.0.10
GitPython==3.1.32
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.24.0
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.25.1
openpyxl==3.1.2
optuna==3.2.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.9.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
rich==13.4.2
rpds-py==0.9.2
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
stack-data==0.6.2
streamlit==1.25.0
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.2.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.20.0
wcwidth==0.2.6
zipp==3.16.2
